Converted to use CreateDIBSection with direct pixel access.
So now it does roughly what the java Swing and WPF test programs
do, but it only uses a constant 1.2MB of memory.  It looks slightly
more jerky than the java one at 60FPS, but I'm wondering if I can
choose to write directly to the DWM backbuffer (I know I've done
that for another win32 test program) to get rid of that.

If it performs so well, it's not clear why it would be worth the
trouble to convert to D3D and custom shaders.